Moved MapView callback handling to NativeBridgeComponent
